Backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
pillow==11.2.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
